Refactoring of directories
