+++ lib/game/components/enemy/enemy_factory.dart
@@ -1,26 +1,20 @@
 import 'package:flame/components.dart';
-import 'package:flutter_game/game/components/crate_enemy.dart';
-import 'package:flutter_game/game/components/imaginary.dart';
+import 'package:flutter_game/game/components/enemy/crate_enemy.dart';
+import 'package:flutter_game/game/components/enemy/imaginary.dart';
 import 'cloud_enemy.dart';
-import 'enemy.dart';
+import 'package:flutter_game/game/components/enemy/enemy.dart';
 
 class EnemyFactory extends BaseComponent {
-  
   Enemy getEnemyType(String type) {
-
-    if(type == "CRATE"){
+    if (type == "CRATE") {
       print("Making Crate");
       return new CrateEnemy();
-    }
-    else if(type == "CLOUD"){
+    } else if (type == "CLOUD") {
       print("Making Cloud");
       return new CloudEnemy();
-    }
-    else{
+    } else {
       print("Making Nothing");
       return new Imaginary();
     }
-
   }
-
 }
