+++ lib/game/components/enemy/imaginary.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_game/game/components/enemy/enemy.dart';
+
+class Imaginary extends Enemy {
+  Imaginary();
+}
