--- lib/game/components/imaginary.dart
@@ -1,5 +1,0 @@
-import 'enemy.dart';
-
-class Imaginary extends Enemy{
-  Imaginary();
-}
