+++ lib/game/components/kiwi.dart
@@ -1,12 +1,11 @@
 import 'package:flame/components.dart';
 import 'package:flame/geometry.dart';
-import 'package:flutter_game/game/components/crate_enemy.dart';
 import 'package:flutter_game/game/game_size_aware.dart';
 import 'package:flutter_game/game/kiwi_game.dart';
 import 'package:flutter_game/game/overlay/end_game_menu.dart';
 import 'package:flutter_game/game/overlay/pause_button.dart';
 
-import 'enemy.dart';
+import 'package:flutter_game/game/components/enemy/enemy.dart';
 
 class Kiwi extends SpriteComponent
     with GameSizeAware, Hitbox, Collidable, HasGameRef<KiwiGame> {
